requirements.txt
# Document Processing
PyMuPDF>=1.23.0
striprtf>=0.0.26
python-docx>=1.0.0

# Database & Vector Store
psycopg2-binary>=2.9.9
pgvector>=0.2.0

# LangChain Ecosystem (Pinned for Stability)
# Using 0.3.x series which is stable and compatible
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
langchain-openai==0.2.1
langchain-postgres==0.0.12

# Orchestration
langgraph==0.2.20

# Core Dependencies
pydantic>=2.7.4,<2.10.0
numpy<2.0.0

# HTTP & API
requests>=2.25.0
httpx>=0.24.0
fastapi
uvicorn

# Utilities
python-dotenv>=1.0.0
orjson>=3.10.0
tiktoken>=0.7.0